Remove Query and QueryDailyHits models and QueryForm from wagtail.search
