+++ wagtail/contrib/search_promotions/tests.py
@@ -738,38 +738,6 @@ class TestGarbageCollectManagementCommand(TestCase):
         )
 
 
-class TestCopyDailyHitsFromWagtailSearchManagementCommand(TestCase):
-    def run_command(self, **options):
-        output = StringIO()
-        management.call_command(
-            "copy_daily_hits_from_wagtailsearch", stdout=output, **options
-        )
-        output.seek(0)
-        return output
-
-    def test_copy(self):
-        # Create some daily hits in the wagtailsearch.{Query,QueryDailyHits} models
-        from wagtail.search.models import Query as WSQuery
-
-        query = WSQuery.get("test query")
-        query.add_hit(date(2021, 8, 24))
-        query.add_hit(date(2021, 8, 24))
-        query.add_hit(date(2021, 7, 1))
-
-        # Check that nothing magically got inserted into the new query model
-        self.assertFalse(Query.objects.exists())
-
-        # Run the management command
-        self.run_command()
-
-        # Check that the query now exists in the new model
-        new_query = Query.objects.get()
-        self.assertEqual(new_query.query_string, "test query")
-
-        # Check daily hits
-        self.assertEqual(new_query.hits, 3)
-
-
 class TestQueryChooserView(WagtailTestUtils, TestCase):
     def setUp(self):
         self.login()
@@ -797,3 +765,91 @@ class TestQueryChooserView(WagtailTestUtils, TestCase):
         # page numbers out of range should return 404
         response = self.get({"p": 9999})
         self.assertEqual(response.status_code, 404)
+
+
+class TestHitCounter(TestCase):
+    def test_no_hits(self):
+        self.assertEqual(Query.get("Hello").hits, 0)
+
+    def test_hit(self):
+        # Add a hit
+        Query.get("Hello").add_hit()
+
+        # Test
+        self.assertEqual(Query.get("Hello").hits, 1)
+
+    def test_10_hits(self):
+        # Add 10 hits
+        for i in range(10):
+            Query.get("Hello").add_hit()
+
+        # Test
+        self.assertEqual(Query.get("Hello").hits, 10)
+
+
+class TestQueryStringNormalisation(TestCase):
+    def setUp(self):
+        self.query = Query.get("  Hello  World!  ")
+
+    def test_normalisation(self):
+        self.assertEqual(str(self.query), "hello world!")
+
+    def test_equivalent_queries(self):
+        queries = [
+            "  Hello World!",
+            "Hello World!  ",
+            "hello  world!",
+            "  Hello  world!  ",
+        ]
+
+        for query in queries:
+            self.assertEqual(self.query, Query.get(query))
+
+    def test_different_queries(self):
+        queries = [
+            "HelloWorld",
+            "HelloWorld!" "  Hello  World!  ",
+            "Hello",
+        ]
+
+        for query in queries:
+            self.assertNotEqual(self.query, Query.get(query))
+
+
+class TestQueryPopularity(TestCase):
+    def test_query_popularity(self):
+        # Add 3 hits to unpopular query
+        for i in range(3):
+            Query.get("unpopular query").add_hit()
+
+        # Add 10 hits to popular query
+        for i in range(10):
+            Query.get("popular query").add_hit()
+
+        # Get most popular queries
+        popular_queries = Query.get_most_popular()
+
+        # Check list
+        self.assertEqual(popular_queries.count(), 2)
+        self.assertEqual(popular_queries[0], Query.get("popular query"))
+        self.assertEqual(popular_queries[1], Query.get("unpopular query"))
+
+        # Add 5 hits to little popular query
+        for i in range(5):
+            Query.get("little popular query").add_hit()
+
+        # Check list again, little popular query should be in the middle
+        self.assertEqual(popular_queries.count(), 3)
+        self.assertEqual(popular_queries[0], Query.get("popular query"))
+        self.assertEqual(popular_queries[1], Query.get("little popular query"))
+        self.assertEqual(popular_queries[2], Query.get("unpopular query"))
+
+        # Unpopular query goes viral!
+        for i in range(20):
+            Query.get("unpopular query").add_hit()
+
+        # Unpopular query should be most popular now
+        self.assertEqual(popular_queries.count(), 3)
+        self.assertEqual(popular_queries[0], Query.get("unpopular query"))
+        self.assertEqual(popular_queries[1], Query.get("popular query"))
+        self.assertEqual(popular_queries[2], Query.get("little popular query"))
