+++ wagtail/search/models.py
@@ -1,9 +1,5 @@
-import datetime
-import warnings
-
 from django import VERSION as DJANGO_VERSION
 from django.apps import apps
-from django.conf import settings
 from django.contrib.contenttypes.fields import GenericForeignKey, GenericRelation
 from django.contrib.contenttypes.models import ContentType
 from django.db import connection, models
@@ -11,122 +7,10 @@ from django.db.models.fields import TextField
 from django.db.models.fields.related import OneToOneField
 from django.db.models.functions import Cast
 from django.db.models.sql.where import WhereNode
-from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
-
-from wagtail.search.utils import MAX_QUERY_STRING_LENGTH, normalise_query_string
-from wagtail.utils.deprecation import RemovedInWagtail60Warning
 
 from .index import class_is_indexed
 from .utils import get_descendants_content_types_pks
-
-
-class Query(models.Model):
-    """
-    Deprecated - This model has been moved to `wagtail.contrib.search_promotions`.
-    This legacy model will be removed in Wagtail 6.0
-    """
-
-    query_string = models.CharField(max_length=MAX_QUERY_STRING_LENGTH, unique=True)
-
-    wagtail_reference_index_ignore = True
-
-    def save(self, *args, **kwargs):
-        # Normalise query string
-        self.query_string = normalise_query_string(self.query_string)
-
-        super().save(*args, **kwargs)
-
-    def add_hit(self, date=None):
-        warnings.warn(
-            "The wagtailsearch.Query model has been moved to wagtail.contrib.search_promotions. "
-            "Please update your code to use the Query model from that app instead.",
-            RemovedInWagtail60Warning,
-            stacklevel=2,
-        )
-        if date is None:
-            date = timezone.now().date()
-        daily_hits, created = QueryDailyHits.objects.get_or_create(
-            query=self, date=date
-        )
-        daily_hits.hits = models.F("hits") + 1
-        daily_hits.save()
-
-    def __str__(self):
-        return self.query_string
-
-    @property
-    def hits(self):
-        hits = self.daily_hits.aggregate(models.Sum("hits"))["hits__sum"]
-        return hits if hits else 0
-
-    @classmethod
-    def garbage_collect(cls):
-        """
-        Deletes all Query records that have no daily hits or editors picks
-        """
-        extra_filter_kwargs = (
-            {
-                "editors_picks__isnull": True,
-            }
-            if hasattr(cls, "editors_picks")
-            else {}
-        )
-        cls.objects.filter(daily_hits__isnull=True, **extra_filter_kwargs).delete()
-
-    @classmethod
-    def get(cls, query_string):
-        return cls.objects.get_or_create(
-            query_string=normalise_query_string(query_string)
-        )[0]
-
-    @classmethod
-    def get_most_popular(cls, date_since=None):
-        # TODO: Implement date_since
-        return (
-            cls.objects.filter(daily_hits__isnull=False)
-            .annotate(_hits=models.Sum("daily_hits__hits"))
-            .distinct()
-            .order_by("-_hits")
-        )
-
-
-class QueryDailyHits(models.Model):
-    """
-    Deprecated - This model has been moved to `wagtail.contrib.search_promotions`.
-    This legacy model will be removed in Wagtail 6.0
-    """
-
-    query = models.ForeignKey(
-        Query, db_index=True, related_name="daily_hits", on_delete=models.CASCADE
-    )
-    date = models.DateField()
-    hits = models.IntegerField(default=0)
-
-    @classmethod
-    def garbage_collect(cls, days=None):
-        """
-        Deletes all QueryDailyHits records that are older than a set number of days
-        """
-
-        warnings.warn(
-            "The wagtailsearch.QueryDailyHits model has been moved to wagtail.contrib.search_promotions. "
-            "Please update your code to use the Query model from that app instead.",
-            RemovedInWagtail60Warning,
-            stacklevel=2,
-        )
-
-        days = (
-            getattr(settings, "WAGTAILSEARCH_HITS_MAX_AGE", 7) if days is None else days
-        )
-        min_date = timezone.now().date() - datetime.timedelta(days)
-
-        cls.objects.filter(date__lt=min_date).delete()
-
-    class Meta:
-        unique_together = (("query", "date"),)
-        verbose_name = _("Query Daily Hits")
-        verbose_name_plural = _("Query Daily Hits")
 
 
 class TextIDGenericRelation(GenericRelation):
